refactor: move AgendaDetailsScreenEvent to separate file
